Move types to seperate file and minimize exports
