+++ packages/runtime/container-runtime/src/pendingStateManager.ts
@@ -11,7 +11,7 @@ import { ICriticalContainerError } from "@fluidframework/container-definitions";
 import { ISequencedDocumentMessage } from "@fluidframework/protocol-definitions";
 import { DataProcessingError, ITelemetryLoggerExt } from "@fluidframework/telemetry-utils";
 
-import { ContainerMessageType, ContainerRuntimeMessage } from "./containerRuntime";
+import { ContainerMessageType, TypedContainerRuntimeMessage } from "./messageTypes";
 import { pkgVersion } from "./packageVersion";
 import { IBatchMetadata } from "./metadata";
 
@@ -243,7 +243,7 @@ export class PendingStateManager implements IDisposable {
 	 * @param message - The message that got ack'd and needs to be processed.
 	 */
 	public processPendingLocalMessage(
-		message: ISequencedDocumentMessage & ContainerRuntimeMessage<string, unknown>,
+		message: ISequencedDocumentMessage & TypedContainerRuntimeMessage<string, unknown>,
 	): unknown {
 		// Pre-processing part - This may be the start of a batch.
 		this.maybeProcessBatchBegin(message);
