+++ packages/runtime/container-runtime/src/test/opLifecycle/outbox.spec.ts
@@ -28,10 +28,10 @@ import {
 } from "../../opLifecycle";
 import {
 	CompressionAlgorithms,
-	ContainerMessageType,
 	ICompressionRuntimeOptions,
 	makeLegacySendBatchFn,
 } from "../../containerRuntime";
+import { ContainerMessageType } from "../../messageTypes";
 
 describe("Outbox", () => {
 	const maxBatchSizeInBytes = 1024;
